Move Processor to entrypoint
